+++ test/sql/install_psql.sql
@@ -1,15 +1,16 @@
 -- install_psql.sql : install E-Maj with a psql script
 --
 -----------------------------
--- install dblink and btree_gist
------------------------------
-CREATE EXTENSION IF NOT EXISTS dblink;
-CREATE EXTENSION IF NOT EXISTS btree_gist;
-
------------------------------
--- emaj installation as extension
+-- emaj installation (not as extension)
 -----------------------------
 \i sql/emaj-devel.sql
+
+-- Test the dblink extension lack (this cannot be easily simulated in the verify.sql unit tests script when emaj is installed as an extension)
+begin;
+  drop extension dblink;
+  select * from emaj.emaj_verify_all();
+  rollback;
+end;
 
 -----------------------------
 -- check installation